Add context menus to many more tables
Fixes #674, and also for other views. In the next step one
could potentially add the shortcuts (usually Edit at C-e) to
more views as well, if that is useful at all.
