+++ name.abuchen.portfolio.ui/src/name/abuchen/portfolio/ui/views/taxonomy/ReBalancingViewer.java
@@ -23,6 +23,7 @@ import name.abuchen.portfolio.money.Values;
 import name.abuchen.portfolio.snapshot.AssetPosition;
 import name.abuchen.portfolio.ui.Images;
 import name.abuchen.portfolio.ui.Messages;
+import name.abuchen.portfolio.ui.editor.AbstractFinanceView;
 import name.abuchen.portfolio.ui.util.Colors;
 import name.abuchen.portfolio.ui.util.SimpleAction;
 import name.abuchen.portfolio.ui.util.swt.ActiveShell;
@@ -34,9 +35,9 @@ import name.abuchen.portfolio.ui.util.viewers.ValueEditingSupport;
 public class ReBalancingViewer extends AbstractNodeTreeViewer
 {
     @Inject
-    public ReBalancingViewer(TaxonomyModel model, TaxonomyNodeRenderer renderer)
+    public ReBalancingViewer(AbstractFinanceView view, TaxonomyModel model, TaxonomyNodeRenderer renderer)
     {
-        super(model, renderer);
+        super(view, model, renderer);
     }
 
     @Override
